+++ packages/Panel/src/components/Dialog/Dialog.styles.js
@@ -1,10 +1,11 @@
 import stylers from "@paprika/stylers";
 import tokens from "@paprika/tokens";
 import styled, { css, keyframes } from "styled-components";
+import * as types from "../../types";
 
-const childPanelCss = ({ isCompact, groupOffsetY, offsetY }) => {
+const childPanelCss = ({ isCompact, groupOffsetY, offset }) => {
   const totalOffset = isCompact ? 48 : 64;
-  const childBottomOffsetY = totalOffset + groupOffsetY + offsetY;
+  const childBottomOffsetY = totalOffset + groupOffsetY + offset.top;
   return css`
     border-bottom-left-radius: ${tokens.spaceSm};
     border-top-left-radius: ${tokens.spaceSm};
@@ -39,11 +40,63 @@ const compactStyles = css`
   padding: ${stylers.spacer(2)};
 `;
 
+function getSlideOutTransform(slideFrom) {
+  switch (slideFrom) {
+    case types.slideFroms.LEFT:
+      return "translateX(-100%)";
+    case types.slideFroms.RIGHT:
+      return "translateX(100%)";
+    case types.slideFroms.BOTTOM:
+      return "translateY(100%)";
+    default:
+      return null;
+  }
+}
+
+function getSlideInTransform(slideFrom) {
+  switch (slideFrom) {
+    case types.slideFroms.LEFT:
+    case types.slideFroms.RIGHT:
+      return "translateX(0)";
+    case types.slideFroms.BOTTOM:
+      return "translateY(0)";
+    default:
+      return null;
+  }
+}
+
 export const Dialog = styled.div(
-  ({ width, isAnimating, isSlideFromLeft, isOpen, hasPushedElement, kind, offsetY, zIndex, isInline }) => {
-    const widthAsPx = Number.isNaN(Number(width)) ? width : `${width}px`;
-    const slideOutTransform = isSlideFromLeft ? "translateX(-100%)" : `translateX(100%)`;
-    const slideInTransform = "translateX(0)";
+  ({ hasPushedElement, height, isAnimating, isInline, isOpen, kind, offset, slideFrom, width, zIndex }) => {
+    const slideOutTransform = getSlideOutTransform(slideFrom);
+    const slideInTransform = getSlideInTransform(slideFrom);
+
+    let calculatedHeight = null;
+    let calculatedWidth = null;
+    let left = null;
+    let right = null;
+    let top = null;
+    let bottom = null;
+    let borderTop = null;
+
+    switch (slideFrom) {
+      case types.slideFroms.LEFT:
+      case types.slideFroms.RIGHT:
+        calculatedHeight = offset.top ? `calc(100% - ${offset.top}px);` : "100%";
+        calculatedWidth = Number.isNaN(Number(width)) ? width : `${width}px`;
+        left = slideFrom === types.slideFroms.LEFT ? 0 : null;
+        right = slideFrom === types.slideFroms.RIGHT ? 0 : null;
+        top = `${offset.top}px`;
+        break;
+      case types.slideFroms.BOTTOM:
+        calculatedHeight = Number.isNaN(Number(height)) ? height : `${height}px`;
+        calculatedWidth = `calc(100% - ${offset.left}px - ${offset.right}px)`;
+        left = `${offset.left}px`;
+        right = `${offset.right}px`;
+        bottom = 0;
+        borderTop = `1px solid ${tokens.border.color}`;
+        break;
+      default:
+    }
 
     const animationStyle = isAnimating
       ? css`
@@ -61,22 +114,22 @@ export const Dialog = styled.div(
       ${animationStyle};
       background: ${tokens.color.white};
       border: 0;
+      border-top: ${borderTop};
+      bottom: ${bottom};
       box-shadow: ${boxShadow};
       box-sizing: border-box;
       ${childPanel}
       display: flex;
       flex-direction: column;
-      height: 100%;
-      ${offsetY ? `height: calc(100% - ${offsetY}px);` : ""}
-      ${isSlideFromLeft ? `left: 0;` : `right: 0;`}
+      height: ${calculatedHeight};
+      left: ${left};
       margin: 0;
       overflow: auto;
       padding: 0;
       ${isInline ? "position: relative;" : "position: fixed;"}
-      right: 0;
-      top: 0;
-      top: ${offsetY}px;
-      width: ${widthAsPx};
+      right: ${right};
+      top: ${top};
+      width: ${calculatedWidth};
       z-index: ${zIndex};
 
       &:focus {
@@ -87,7 +140,7 @@ export const Dialog = styled.div(
 );
 
 export const DialogContent = styled.div(
-  ({ hasPushedElement, isSlideFromLeft, isCompact, kind }) => css`
+  ({ hasPushedElement, slideFrom, isCompact, kind }) => css`
   flex-grow: 1;
   margin: 0;
   padding: 0;
@@ -96,8 +149,8 @@ export const DialogContent = styled.div(
     ${stylers.focusRing.subtle(true)};
   }
 
-  ${hasPushedElement && isSlideFromLeft ? `border-right: 1px solid ${tokens.border.color}` : ""}
-  ${hasPushedElement && !isSlideFromLeft ? `border-left: 1px solid ${tokens.border.color}` : ""}
+  ${hasPushedElement && slideFrom === types.slideFroms.LEFT ? `border-right: 1px solid ${tokens.border.color}` : ""}
+  ${hasPushedElement && slideFrom === types.slideFroms.RIGHT ? `border-left: 1px solid ${tokens.border.color}` : ""}
   ${isCompact || kind === "child" ? compactStyles : ""};
 `
 );